pyproject.toml
[project]
name = "hyperbolic-medical"
version = "0.1.0"
description = "Hyperbolic domain generalization for medical imaging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "gdown>=5.2.0",
    "geoopt>=0.5.1",
    "medmnist>=3.0.1",
    "medmnistc>=0.1.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "timm>=1.0.15",
    "tqdm>=4.67.1",
    "wandb>=0.20.1",
    "wilds>=2.0.0",
]

[project.scripts]
hypcbc-train = "hypcbc.main:main"
hypcbc-preprocess = "hypcbc.preprocessing:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hypcbc"]
